k/record_batcher: Move to kafka/data
Followup to https://github.com/redpanda-data/redpanda/pull/23775
