+++ src/v/kafka/client/test/utils.h
@@ -23,7 +23,3 @@ inline model::record_batch make_batch(model::offset offset, size_t count) {
     }
     return std::move(builder).build();
 }
-
-namespace kafka::client::testing {
-iobuf random_length_iobuf(size_t data_max);
-} // namespace kafka::client::testing
