+++ src/v/kafka/data/BUILD
@@ -41,3 +41,28 @@ redpanda_cc_library(
         "@seastar",
     ],
 )
+
+redpanda_cc_library(
+    name = "record_batcher",
+    srcs = [
+        "record_batcher.cc",
+    ],
+    hdrs = [
+        "record_batcher.h",
+    ],
+    implementation_deps = [
+        ":logger",
+        "//src/v/storage:record_batch_builder",
+        "//src/v/utils:human",
+        "//src/v/utils:truncating_logger",
+        "@abseil-cpp//absl/algorithm:container",
+    ],
+    include_prefix = "kafka/data",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/v/base",
+        "//src/v/bytes:iobuf",
+        "//src/v/model",
+        "@seastar",
+    ],
+)
