+++ src/v/kafka/data/CMakeLists.txt
@@ -1,7 +1,16 @@
 v_cc_library(
-  NAME kafka_partition_proxy
+  NAME kd_logger
+  HDRS
+    logger.h
   SRCS
     logger.cc
+  DEPS
+    v::base
+)
+
+v_cc_library(
+  NAME kafka_partition_proxy
+  SRCS
     partition_proxy.cc
     replicated_partition.cc
   DEPS
@@ -9,5 +18,18 @@ v_cc_library(
     v::cloud_storage
     v::cluster
     v::kafka_protocol
+    v::kd_logger
     v::model
+)
+
+v_cc_library(
+  NAME kd_record_batcher
+  HDRS
+    record_batcher.h
+  SRCS
+    record_batcher.cc
+  DEPS
+    v::kd_logger
+    v::model
+    v::storage
 )
