+++ src/v/kafka/data/test/utils.h
@@ -1,0 +1,18 @@
+/*
+ * Copyright 2020 Redpanda Data, Inc.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the file licenses/BSL.md
+ *
+ * As of the Change Date specified in that file, in accordance with
+ * the Business Source License, use of this software will be governed
+ * by the Apache License, Version 2.0
+ */
+
+#pragma once
+
+#include "bytes/iobuf.h"
+
+namespace kafka::data::testing {
+iobuf random_length_iobuf(size_t data_max);
+} // namespace kafka::data::testing
