+++ src/v/security/audit/BUILD
@@ -36,7 +36,7 @@ redpanda_cc_library(
         "types.h",
     ],
     implementation_deps = [
-        "//src/v/kafka/client:record_batcher",
+        "//src/v/kafka/data:record_batcher",
         "@abseil-cpp//absl/algorithm:container",
     ],
     include_prefix = "security/audit",
