+++ src/v/security/audit/CMakeLists.txt
@@ -12,7 +12,7 @@ v_cc_library(
     v::utils
     v::config
     v::cluster
-    v::kc_record_batcher
+    v::kd_record_batcher
 )
 
 add_subdirectory(schemas)
