+++ src/v/transform/logging/rpc_client.cc
@@ -13,7 +13,7 @@
 
 #include "cluster/metadata_cache.h"
 #include "hashing/murmur.h"
-#include "kafka/client/record_batcher.h"
+#include "kafka/data/record_batcher.h"
 #include "logger.h"
 #include "model/namespace.h"
 
@@ -41,7 +41,7 @@ rpc_client::write(model::partition_id pid, io::json_batches batches) {
     auto max_batch_size = cfg->properties.batch_max_bytes.value_or(
       config::shard_local_cfg().kafka_batch_max_bytes());
 
-    kafka::client::record_batcher batcher{max_batch_size, &tlg_log};
+    kafka::data::record_batcher batcher{max_batch_size, &tlg_log};
 
     while (!batches.empty()) {
         auto json_batch = std::move(batches.front());
